requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==24.1.0
aiogram==2.15
aiohttp==3.9.5
aiohttp-retry==2.9.1
aiormq==6.4.2
aiosignal==1.3.1
annotated-doc==0.0.4
anyio==4.5.2
APScheduler==3.9.1.post1
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
babel==2.17.0
backoff==2.2.1
backports.zoneinfo==0.2.1
bcrypt==5.0.0
bidict==0.23.1
boto3==1.37.38
botocore==1.37.38
build==1.2.2.post1
CacheControl==0.12.14
cachetools==5.5.2
certifi==2026.1.4
cffi==1.17.1
charset-normalizer==3.4.4
chroma-hnswlib==0.7.6
click==8.1.8
cloudpickle==3.1.2
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confluent-kafka==2.12.2
cryptography==46.0.3
cycler==0.12.1
dask==2022.10.2
Deprecated==1.3.1
distro==1.9.0
dnspython==2.3.0
docopt==0.6.2
durationpy==0.10
exceptiongroup==1.3.1
fastapi==0.124.4
fbmessenger==6.0.0
filelock==3.16.1
fire==0.7.1
flatbuffers==25.12.19
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
gast==0.4.0
google-ai-generativelanguage==0.1.0
google-api-core==2.29.0
google-auth==2.45.0
google-auth-oauthlib==1.0.0
google-generativeai==0.1.0rc1
google-pasta==0.2.0
googleapis-common-protos==1.72.0
graphlib_backport==1.1.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.62.3
h11==0.16.0
h5py==3.11.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
jax==0.4.13
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonpickle==3.0.4
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.7
kubernetes==35.0.0
libclang==18.1.1
locket==1.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mattermostwrapper==2.2
mdurl==0.1.2
ml-dtypes==0.2.0
mmh3==5.0.1
mpmath==1.3.0
msgpack==1.1.1
multidict==5.2.0
networkx==2.6.3
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
onnxruntime==1.19.2
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
opt_einsum==3.4.0
orjson==3.10.15
overrides==7.7.0
packaging==20.9
pamqp==3.2.1
partd==1.4.1
pillow==10.4.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
pluggy==1.5.0
portalocker==2.10.1
posthog==4.2.0
prompt-toolkit==3.0.28
propcache==0.2.0
proto-plus==1.27.1
protobuf==4.23.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==1.10.9
pydot==1.4.2
Pygments==2.19.2
PyJWT==2.9.0
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.1.4
PyPika==0.51.1
pyproject_hooks==1.2.0
pyrsistent==0.20.0
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.13.0
python-socketio==5.16.0
pytz==2022.7.1
PyYAML==6.0.3
questionary==1.10.0
randomname==0.1.5
rasa==3.6.21
rasa-sdk==3.6.2
redis==4.6.0
regex==2022.10.31
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rocketchat-API==1.30.0
rsa==4.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
s3transfer==0.11.5
safetensors==0.4.5
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.10.1
sentry-sdk==1.14.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sklearn-crfsuite==0.3.6
skops==0.9.0
slack_sdk==3.39.0
sniffio==1.3.1
SQLAlchemy==1.4.54
starlette==0.44.0
structlog==23.3.0
structlog-sentry==2.1.0
sympy==1.13.3
tabulate==0.9.0
tarsafe==0.0.4
tenacity==9.0.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-text==2.12.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tokenizers==0.20.3
tomli==2.4.0
toolz==1.0.0
torch==2.4.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
twilio==8.2.2
typer==0.20.1
typing-utils==0.1.0
typing_extensions==4.13.2
tzdata==2025.3
tzlocal==5.2
ujson==5.10.0
urllib3==1.26.20
uvicorn==0.33.0
uvloop==0.22.1
watchfiles==0.24.0
wcwidth==0.2.14
webexteamssdk==1.6.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.6
wrapt==1.14.2
wsproto==1.2.0
yarl==1.15.2
zipp==3.20.2
